Move logic to move players out of the arena module
