+++ src/components/arena.ts
@@ -1,17 +1,15 @@
 import { Player, PLAYER_RADIUS } from '../player'
 import { Shot } from '../shot'
 import { Mine } from '../mine'
-import { Movement } from '../messages'
 import { RadarScan, ScanResult } from './radar'
 import { Position } from '../types'
+import { round } from '../helpers'
 import asyncStateUpdate from '../domain/async-state-update'
 
 export type Success<T = {}> = { status: 'ok' } & T
 export type Failure<T> = { status: 'ko' } & T
+// TODO move this to Types
 export type Result<T, F> = Success<T> | Failure<F>
-
-// TODO replace process.env with a configuration value passed to the app
-const MOVEMENT_SPEED = process.env.MOVEMENT_SPEED ? Number(process.env.MOVEMENT_SPEED) : 1
 
 export function asSuccess<T, F> (result: Result<T, F>): Success<T> | never {
   if (result.status === 'ok') {
@@ -20,15 +18,6 @@ export function asSuccess<T, F> (result: Result<T, F>): Success<T> | never {
 
   // TODO include the failure details in the thrown error
   throw new Error('Expected a Success got a Failure')
-}
-
-/*
- * Round with 5 decimal digits
- */
-function round (value: number): number {
-  const multiplier = Math.pow(10, 5)
-
-  return Math.round(multiplier * value) / multiplier
 }
 
 // TODO maybe rename this type and ArenaShot to
@@ -254,63 +243,10 @@ export class Arena {
     return update.updates
   }
 
-  // TODO this should take an ArenaPlayer or instead the id of the player to move
-  movePlayer (movement: Movement, player: Player): Success<{ position: Position }> {
-    const arenaPlayer = this.arenaPlayers.find((arenaPlayer) => arenaPlayer.id === player.id)
-    // TODO throw if the player hasn't been found
-    const newPosition = this.calculateNewPlayerPosition(movement, player, arenaPlayer!.position)
-
-    const { x, y } = newPosition
-    const collides = this.arenaPlayers.find((arenaPlayer) => {
-      if (arenaPlayer.id === player.id) {
-        return false
-      }
-
-      /*
-       * Formula got at http://stackoverflow.com/a/8367547/1078859
-       * (R0-R1)^2 <= (x0-x1)^2+(y0-y1)^2 <= (R0+R1)^2
-       */
-
-      const { x: ox, y: oy } = arenaPlayer.position
-      const value = Math.pow((x - ox), 2) + Math.pow((y - oy), 2)
-      // TODO the Math.pow(PLAYER_RADIUS - PLAYER_RADIUS, 2) part is useless
-      // but will it keep it here in case we have players with different radius
-
-      return Math.pow(PLAYER_RADIUS - PLAYER_RADIUS, 2) <= value && value <= Math.pow(PLAYER_RADIUS + PLAYER_RADIUS, 2)
-    })
-
-    if (this.isPlayerPositionWithinBoundaries(newPosition) && collides === undefined) {
-      arenaPlayer!.position = newPosition
-
-      return {
-        status: 'ok',
-        position: newPosition
-      }
-    } else {
-      return {
-        status: 'ok',
-        position: arenaPlayer!.position
-      }
-    }
-  }
-
   rotatePlayer (rotation: number, player: Player): void {
     const arenaPlayer = this.arenaPlayers.find((arenaPlayer) => arenaPlayer.id === player.id)
     // TODO throw if the player is not found
     arenaPlayer!.rotation = rotation
-  }
-
-  private calculateNewPlayerPosition (movement: Movement, player: Player, currentPosition: Position): Position {
-    const direction = movement.direction === 'forward' ? 1 : -1
-    // TODO read the movement speed from the configuration
-    const magnitude = direction * MOVEMENT_SPEED
-    const radians = (player.rotation * Math.PI) / 180
-    const dX = magnitude * Math.cos(radians)
-    const dY = magnitude * Math.sin(radians)
-    const newX = round(dX + currentPosition.x)
-    const newY = round(dY + currentPosition.y)
-
-    return { x: newX, y: newY }
   }
 
   private isPlayerPositionWithinBoundaries ({ x, y }: Position): boolean {
